🐛 Исправлена ошибка типа данных в togglePin
Проблема: PostgreSQL требовал явного указания типа для NULL в CASE WHEN
Решение: Добавлены ::INTEGER приведения типов в SQL запросе

До: pinned_by = CASE WHEN is_pinned = false THEN $2 ELSE NULL END
После: pinned_by = CASE WHEN is_pinned = false THEN $2::INTEGER ELSE NULL::INTEGER END

Теперь закрепление/открепление сообщений работает корректно

+++ backend/models/Message.js
@@ -128,7 +128,7 @@ class Message {
         `UPDATE messages 
          SET 
            is_pinned = NOT is_pinned,
-           pinned_by = CASE WHEN is_pinned = false THEN $2 ELSE NULL END,
+           pinned_by = CASE WHEN is_pinned = false THEN $2::INTEGER ELSE NULL::INTEGER END,
            pinned_at = CASE WHEN is_pinned = false THEN CURRENT_TIMESTAMP ELSE NULL END
          WHERE id = $1
          RETURNING *`,
